requirements.txt
appdirs==1.4.3
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
Django==1.11.6
future==0.16.0
gmusicapi==10.1.2
gpsoauth==0.4.1
httplib2==0.10.3
idna==2.6
MechanicalSoup==0.8.0
mock==2.0.0
mutagen==1.38
oauth2client==4.1.2
pbr==3.1.1
proboscis==1.2.6.0
protobuf==3.4.0
pyasn1==0.3.7
pyasn1-modules==0.1.4
pycryptodomex==3.4.7
pyglet==1.2.4
python-dateutil==2.6.1
python-vlc==1.1.2
pytz==2017.2
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
validictory==1.1.1
